backend/requirements.txt
anyio==3.5.0
asgiref==3.5.1
bcrypt==3.2.2
boto3==1.22.10
botocore==1.25.10
certifi==2024.2.2
cffi==1.15.0
click==8.1.3
colorama==0.4.4
distro==1.9.0
fastapi==0.77.1
greenlet==1.1.2
h11==0.13.0
httpcore==1.0.5
httpx==0.27.0
idna==3.3
jmespath==1.0.0
openai==1.16.2
passlib==1.7.4
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==1.26.9
uvicorn==0.17.6
